+++ GeneralsMD/Code/GameEngine/Source/GameClient/GUI/Shell/Shell.cpp
@@ -514,14 +514,30 @@ void Shell::showShell( Bool runInit )
 
 void Shell::showShellMap(Bool useShellMap )
 {
+	printf("SHELLMAP: showShellMap(%d) shellMapOn=%d initialFile='%s' gameLogic=%p\n",
+		(int)useShellMap, (int)TheGlobalData->m_shellMapOn,
+		TheGlobalData->m_initialFile.str(), (void*)TheGameLogic);
+	fflush(stdout);
+
 	// we don't want any of this to show if we're loading straight into a file
 	if (TheGlobalData->m_initialFile.isNotEmpty() || !TheGameLogic || !TheGlobalData->m_simulateReplays.empty())
+	{
+		printf("SHELLMAP: early return\n"); fflush(stdout);
 		return;
+	}
 	if(useShellMap && TheGlobalData->m_shellMapOn)
 	{
+		printf("SHELLMAP: shellMap path, isInGame=%d gameMode=%d shellMapName='%s'\n",
+			(int)TheGameLogic->isInGame(), (int)TheGameLogic->getGameMode(),
+			TheGlobalData->m_shellMapName.str());
+		fflush(stdout);
+
 		// we're already in a shell game, return
 		if(TheGameLogic->isInGame() && TheGameLogic->getGameMode() == GAME_SHELL)
+		{
+			printf("SHELLMAP: already in shell game, return\n"); fflush(stdout);
 			return;
+		}
 		// we're in some other kind of game, clear it out foo!
 		if(TheGameLogic->isInGame())
 			TheGameLogic->exitGame();
@@ -531,6 +547,9 @@ void Shell::showShellMap(Bool useShellMap )
 		GameMessage *msg = TheMessageStream->appendMessage( GameMessage::MSG_NEW_GAME );
 		msg->appendIntegerArgument(GAME_SHELL);
 		m_shellMapOn = TRUE;
+		printf("SHELLMAP: MSG_NEW_GAME sent, pendingFile='%s'\n",
+			TheGlobalData->m_pendingFile.str());
+		fflush(stdout);
 	}
 	else
 	{